Validaciones por si no hay proyectos/productos/investigadores

+++ resources/views/admin-invest/products/index.blade.php
@@ -9,42 +9,47 @@
                     Listado de Productos
                     <a href="{{ route('products.create') }}" class="btn btn-sm btn-success float-right">Crear Producto</a>
                 </div>
+                
+                @if ($products->items() != null)
+                    <div class="card-body">
+                        <table class="table table-striped table-hover">
+                            <thead>
+                                <tr>
+                                    <th width="10px">ID</th>
+                                    <th style="width: 5cm">Nombre</th>
+                                    <th>Descripcion</th>
+                                    <th colspan="2">&nbsp;</th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                @foreach($products as $product)
+                                <tr>
 
-                <div class="card-body">
-                    <table class="table table-striped table-hover">
-                        <thead>
-                            <tr>
-                                <th width="10px">ID</th>
-                                <th style="width: 5cm">Nombre</th>
-                                <th>Descripcion</th>
-                                <th colspan="2">&nbsp;</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            @foreach($products as $product)
-                            <tr>
+                                    <td>{{ $product->id}}</td>
+                                    <td>{{ $product->name}}</td>
+                                    <td>{{$product->description}}</td>
+                                    <td width="10px">
+                                        <a href=" {{route('products.edit', $product->id) }}" class="btn btn-sm btn-secondary">
+                                            Editar
+                                        </a>
+                                    </td>
+                                    <td width="10px">
+                                        {!! Form::open(['route' => ['products.destroy', $product->id], 'method' => 'DELETE']) !!}
+                                            <button class="btn btn-sm btn-danger">
+                                                Eliminar
+                                            </button>
+                                        {!! Form::close() !!}
+                                    </td>
+                                </tr>
+                                @endforeach
+                            </tbody>
+                        </table>
+                        {{ $products->render()}}
+                    </div>
+                @else
+                    <p class="display-4 text-center"> Aún no se registran productos </p>       
+                @endif
 
-                                <td>{{ $product->id}}</td>
-                                <td>{{ $product->name}}</td>
-                                <td>{{$product->description}}</td>
-                                <td width="10px">
-                                    <a href=" {{route('products.edit', $product->id) }}" class="btn btn-sm btn-secondary">
-                                        Editar
-                                    </a>
-                                </td>
-                                <td width="10px">
-                                    {!! Form::open(['route' => ['products.destroy', $product->id], 'method' => 'DELETE']) !!}
-                                        <button class="btn btn-sm btn-danger">
-                                            Eliminar
-                                        </button>
-                                    {!! Form::close() !!}
-                                </td>
-                            </tr>
-                            @endforeach
-                        </tbody>
-                    </table>
-                    {{ $products->render()}}
-                </div>
             </div>
         </div>
     </div>
